// types.ts
export enum UserRole {
  USER = "USER",
  ADMIN = "ADMIN",
}

export type User = {
  id: string;
  name: string;
  email: string;
  emailVerified?: Date | null;
  image?: string | null;
  password?: string | null;
  role: UserRole;
  accounts: Account[];
  access_token?: string | null;
};

export type Account = {
  id: string;
  userId: string;
  type: string;
  provider: string;
  providerAccountId: string;
  refresh_token?: string | null;
  access_token?: string | null;
  expires_at?: number | null;
  token_type?: string | null;
  scope?: string | null;
  id_token?: string | null;
  session_state?: string | null;
  user: User;
};